% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_map.R
\name{map_ui}
\alias{map_ui}
\title{Leaflet Map UI Module}
\usage{
map_ui(id)
}
\arguments{
\item{id}{A `character` string. The namespace ID for the module.}
}
\value{
A `shiny.tag` object with the leaflet map output UI.
}
\description{
Creates the user interface (UI) for a leaflet map. This
  function is intended to be used as a Shiny module.
}
