% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_selector.R
\name{selector_ui}
\alias{selector_ui}
\title{Tag Selector UI Module (Selectize Version)}
\usage{
selector_ui(id)
}
\arguments{
\item{id}{A `character` string. The namespace ID for the module.}
}
\value{
A `shiny.tag` object with the selectize input UI.
}
\description{
Creates a UI component using `shiny::selectizeInput` for selecting multiple
tags.
}
\note{
The choices for this input are initialized as `NULL` and must be populated
from the server using `shiny::updateSelectizeInput()`. The function also
depends on a globally defined variable, `maximum_selector_items`, to set
the maximum number of selectable items.
}
